Devices/Graphics/HGSMI: warning
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@44110 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Graphics/HGSMI/HGSMIHost.cpp
@@ -1329,7 +1329,7 @@ int HGSMIHostSaveStateExec (PHGSMIINSTANCE pIns, PSSMHANDLE pSSM)
         SSMR3PutU32 (pSSM, HGSMIHeapOffset(&pIns->hostHeap));
         SSMR3PutU32 (pSSM, HGSMIHeapSize(&pIns->hostHeap));
         /* need save mem pointer to calculate offset on restore */
-        SSMR3PutU64 (pSSM, (uint64_t)pIns->area.pu8Base);
+        SSMR3PutU64 (pSSM, (uint64_t)(uintptr_t)pIns->area.pu8Base);
         rc = hgsmiFIFOLock (pIns);
         if(RT_SUCCESS(rc))
         {
